requirements.txt
fastapi
uvicorn
tensorflow
keras
opencv-python-headless
pillow
scikit-learn
